services/scraper-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
playwright==1.44.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
pydantic==2.7.0
